requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0.1
django-constance==3.1.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-money==3.4.1
django-phonenumber-field==7.3.0
django-picklefield==3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
install==1.3.5
oauthlib==3.2.2
phonenumbers==8.13.27
pillow==10.2.0
py-moneyed==3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
